New theme - 'customProperties'
Added new theme that will allow us to use CSS custom properties to control how the data grid is appearing. Inside the 'revo-grid.custom-properties.scss' file, we are using CSS custom properties with the fallback values.
This way, if the properties are defined they will be applied, if not the fallback value will be set.